Fix: Разделение кода на .hpp и .cpp, файл сборки .sh
