+++ rolton-basic.cpp
@@ -3,6 +3,7 @@
 #include <vector>
 #include <string.h>
 #include <cstring>
+
 #include "parser.hpp"
 #include "lexer.hpp"
 #include "compiler.hpp"
@@ -13,7 +14,7 @@ int main(int argc, char** argv) {
     string line;
     vector<rbl_types::ast_type> ast_full;
     vector<rbl_types::var> vars_vector = {{"SYSVERSION", "string", "1.0.1-alpha"}, {"SYSDEVELOPER", "string", "KirillMos1"}};
-    
+
     do {
         cout<<"\n\nEnter code: ";
         getline(cin, line);
@@ -25,7 +26,7 @@ int main(int argc, char** argv) {
         cout<<"[LOG] ADD IN FULL AST\n";
         ast_full.push_back(ast);
     } while (line != "EXIT"); // TODO 0xd001
-    
+
     cout<<"[LOG] START TRANSLATING\n";
     string translated = compiler::translating(ast_full, vars_vector);
     std::cout << "Your name for out file: program";
