+++ types.hpp
@@ -1,12 +1,14 @@
+#pragma once
+
 #include <cstdlib>
 #include <string>
-#ifndef TYPES_H
-#define TYPES_H
+#include <vector>
 
 #define PRINT_WORD tokens[0].type == "WORD" and tokens[0].value == "PRINT"
 #define VARINT_WORD tokens[0].type == "WORD" and tokens[0].value == "VARINT"
 #define VARSTR_WORD tokens[0].type == "WORD" and tokens[0].value == "VARSTR"
 #define TYPE_WORD tokens[0].type == "WORD" and tokens[0].value == "TYPE"
+
 namespace rbl_types {
     typedef struct {
         std::string name;
@@ -18,4 +20,3 @@ namespace rbl_types {
         std::vector<std::string> args;
     } ast_type;
 };
-#endif
